+++ Sources/System/System.cpp
@@ -6,16 +6,16 @@
 //  Copyright © 2017 VladasZ. All rights reserved.
 //
 
-#include "System/Platform.h"
-#include "System/System.hpp"
-#include "Tools/Log.hpp"
-#include "Tools/Debug.hpp"
+#include "Platform.h"
+#include "System.hpp"
+#include "Log.hpp"
+#include "Debug.hpp"
 
 #ifdef APPLE
 #include <unistd.h>
 #endif
 
-void System::sleep(double interval) {
+void System::sleep(Float interval) {
     
 #ifdef APPLE
     usleep(interval * 1000000);
